Move pubsub-connector into src/pubspub-connector
