+++ index.js
@@ -4,7 +4,6 @@ const http = require('http');
 const defaultPort = 80;
 const port = process.env.MS_PORT || defaultPort;
 const app = express();
-const fileStatusUpdate = require("./src/file-status-update");
 const pubsubConnector = require("./src/pubsub-connector");
 const redisPubsub = require("./src/redis-pubsub");
 const restartReboot = require("./src/messages/restart-reboot");
@@ -22,7 +21,7 @@ const gcs = require("./src/version-compare/gcs");
 const querystring = require("querystring");
 const url = require("url");
 
-redisPubsub.init([restartReboot, fileStatusUpdate]);
+redisPubsub.init([restartReboot, pubsubConnector.fileUpdateHandler]);
 
 const primus = new Primus(server, {transformer: 'uws', pathname: 'messaging/primus'});
