+++ src/pubsub-connector/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  postHandler: require("./post-handler.js"),
+  fileUpdateHandler: require("./file-update-handler.js")
+}
