+++ test/unit/file-update-handler.js
@@ -1,6 +1,6 @@
 /* eslint-env mocha */
 const assert = require("assert");
-const fileStatusUpdate = require("../../src/file-status-update");
+const fileUpdateHandler = require("../../src/pubsub-connector/file-update-handler");
 const simple = require("simple-mock");
 const db = require("../../src/db/api");
 const displayConnections = require("../../src/messages/display-connections");
@@ -34,7 +34,7 @@ describe("Pub/sub Update", ()=>{
     simple.restore(db.fileMetadata, "getWatchersFor");
     simple.mock(db.fileMetadata, "getWatchersFor").resolveWith([]);
 
-    return fileStatusUpdate.processUpdate(testIncomingADDMessage)
+    return fileUpdateHandler.processUpdate(testIncomingADDMessage)
     .then(()=>{
       assert.equal(db.fileMetadata.getWatchersFor.callCount, 1);
       assert([
@@ -50,7 +50,7 @@ describe("Pub/sub Update", ()=>{
   it("notifies displays but doesn't change db if the data was shared from a different pod", ()=>{
     const otherPodMsg = Object.assign({}, testIncomingADDMessage, {podname: "other"});
 
-    return fileStatusUpdate.processUpdate(otherPodMsg)
+    return fileUpdateHandler.processUpdate(otherPodMsg)
     .then(()=>{
       assert.equal(db.fileMetadata.getWatchersFor.callCount, 1);
       assert.equal(displayConnections.sendMessage.callCount, watchers.length);
@@ -68,7 +68,7 @@ describe("Pub/sub Update", ()=>{
   it("updates db on DELETE", ()=>{
     const delPodMsg = Object.assign({}, testIncomingADDMessage, {type: "DELETE"});
 
-    return fileStatusUpdate.processUpdate(delPodMsg)
+    return fileUpdateHandler.processUpdate(delPodMsg)
     .then(()=>{
       assert.equal(db.fileMetadata.getWatchersFor.callCount, 1);
       assert.equal(db.fileMetadata.deleteMetadata.callCount, 1);
@@ -85,7 +85,7 @@ describe("Pub/sub Update", ()=>{
   it("updates db on UPDATE", ()=>{
     const updMsg = Object.assign({}, testIncomingADDMessage, {type: "UPDATE"});
 
-    return fileStatusUpdate.processUpdate(updMsg)
+    return fileUpdateHandler.processUpdate(updMsg)
     .then(()=>{
       assert.equal(db.fileMetadata.getWatchersFor.callCount, 1);
       assert.equal(db.fileMetadata.setFileVersion.callCount, 1);
@@ -100,7 +100,7 @@ describe("Pub/sub Update", ()=>{
   });
 
   it("updates db on ADD", ()=>{
-    return fileStatusUpdate.processUpdate(testIncomingADDMessage)
+    return fileUpdateHandler.processUpdate(testIncomingADDMessage)
     .then(()=>{
       assert.equal(db.fileMetadata.getWatchersFor.callCount, 1);
       assert.equal(db.fileMetadata.setFileVersion.callCount, 1);
@@ -116,7 +116,7 @@ describe("Pub/sub Update", ()=>{
 
   it("doesn't update db or send messages if the type is invalid", ()=>{
     const invalidMsg = Object.assign({}, testIncomingADDMessage, {type: "INVALID"});
-    return fileStatusUpdate.processUpdate(invalidMsg)
+    return fileUpdateHandler.processUpdate(invalidMsg)
     .then(()=>{
       assert.equal(db.fileMetadata.getWatchersFor.callCount, 1);
       assert(logger.log.lastCall.arg, "Invalid notification type received");
